Rewrote SequelizeMeta inserts
Using MariaDB/MySQL, the operations on SequelizeMeta fail as the driver does not expect the name SequelizeMeta to be surrounded by quotation marks. I was using this package in my project and noticed the error message kept getting hung up on quotation marks, so I manually removed them from my local copy to get it running.